fix: Filter complete views based on the target.
The usage pattern within iql.query had changed since this was conceived.
Initially we had planned to filter target and conditions for variables.

But that's not what happens. We only have the target - which suffices
for the optimization. But which ended up providing a different set
of column variables from iql.query to iql.inference.